+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Upcoming
 * Mobs are now affected by blindness (They see only 10% of range)
+* Drowned should get stuck less
 * Witches no longer heal themselves if above 70% health
 * Mobs now flee from the warden instead of attaking it
 * Fixed neutral mobs not working